feat: expiration column for myOffers
